完善JSON-RPC接口；修复JSON库中的一些类型错误；

+++ AntSharesCore/IO/Json/JNumber.cs
@@ -60,7 +60,7 @@ namespace AntShares.IO.Json
                 return true;
             if (type.IsEnum && Enum.IsDefined(type, Convert.ChangeType(Value, type.GetEnumUnderlyingType())))
                 return true;
-            if (type == typeof(decimal))
+            if (type == typeof(double))
                 return true;
             if (type == typeof(string))
                 return true;
